requirements.txt
selenium>=4.10.0
requests>=2.31.0
random-user-agent>=1.0.1
Pillow>=10.0.0
